requirements.txt
Pillow==9.5.0
python-dotenv==1.0.0
Pyrogram==2.0.106
